refactor: 重構部分名稱、用法
- revert: Scraper 不再使用 class 包裝
