+++ src/api/routers/resources/careers.py
@@ -1,8 +1,7 @@
 from fastapi import APIRouter
 
-from src.utils.scraper import rpage_scraper
 from src.api import schemas
-
+from src.utils.scrapers import rpage_scraper
 
 router = APIRouter()
 
@@ -12,6 +11,6 @@ async def get_bulletin_recruitment():
     """
     獲取清華公佈欄的徵才公告。
     """
-    return rpage_scraper.announcement(
+    return rpage_scraper.get_announcement(
         "https://bulletin.site.nthu.edu.tw/p/403-1086-5075-1.php?Lang=zh-tw"
     )
