+++ src/utils/scrapers/goodjob_scraper.py
