+++ src/utils/scrapers/newsletter_scraper.py
